+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K6DHS9ZC84NY3NEYVWAN9P31"
+        "01K6G1JW64AHWAV44Y5QA9Y1K1"
     )
     r"""The id of the resource"""