+++ packages/orq-rc/src/orq_ai_sdk/models/streamrunagentop.py
@@ -603,7 +603,7 @@ class RunAgentRequestToolHTTPTool(BaseModel):
     http: RunAgentRequestToolHTTP
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K6DHS9W2ZKYN1YZSHPGCR0MJ"
+        "01K6G1JW2BYTV1S78FW7HKSEKZ"
     )
 
     requires_approval: Optional[bool] = False