+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -119,5 +119,5 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-30T14:52:41.173Z")
+    updated: Optional[datetime] = parse_datetime("2025-10-01T14:07:15.793Z")
     r"""The date and time the resource was last updated"""