+++ packages/orq-rc/RELEASES.md
@@ -3049,3 +3049,13 @@ Based on:
 - [python v3.13.0-rc.52] packages/orq-rc
 ### Releases
 - [PyPI v3.13.0-rc.52] https://pypi.org/project/orq-ai-sdk/3.13.0-rc.52 - packages/orq-rc
+
+## 2025-10-01 14:08:59
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.631.5 (2.716.16) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.14.0-rc.0] packages/orq-rc
+### Releases
+- [PyPI v3.14.0-rc.0] https://pypi.org/project/orq-ai-sdk/3.14.0-rc.0 - packages/orq-rc